refactor(store): extract ssh tunnel domain into a zustand slice
First slice of the appStore.ts monolith (#2077). Moves the SSH tunnel state
(`tunnels`, `tunnelStates`) and its CRUD/lifecycle actions (loadTunnels,
saveTunnel, deleteTunnel, startTunnel, stopTunnel, reconnectTunnel,
updateTunnelState) plus the `_tunnelStartInFlight`/`_tunnelStopInFlight`
re-entrancy guards into a new `StateCreator` slice at
`src/store/slices/tunnelSlice.ts`, combined into the root `create()` via spread.

Pure structural extraction — zero behavior change. Each action still receives
the shared `set`/`get` typed against the full `AppState`, so the public store
shape and every action behave identically; consumers need no changes.
`openTunnelEditorTab` deliberately stays in the root store: it creates a tab via
the module-local `createTab`/`tabCounter` factory and belongs to the panel/tab
domain, moving with that slice later.

`AppState` is now exported and extends `TunnelSlice`. Verified via the full
frontend test suite (4662 tests, incl. the tunnel delete/reentrancy specs),
tsc, eslint, and prettier — all green.
